pyproject.toml
[project]
name = "clicx"
description = ""
readme = "README.md"

[project.scripts]
clicx = "clicx.cli.command:main"

[tool.setuptools]
scripts = ["clicx-bin"]

[tool.poetry]
name = "clicx"
version = "0.1.0"
description = ""
authors = ["martin <mart337i@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
jinja2 = "^3.1.6"
python-dotenv = "^1.0.1"
typer-cli = "^0.15.2"
rich  = "^13.9.4"
fastapi = {extras = ["standard"], version = "^0.115.11"}
urllib3 = "^2.3.0"
psutil = "^7.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

